requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.10.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
deprecation==2.1.0
django==5.2.7
django-cors-headers==4.9.0
django-filter==25.2
django-redis==6.0.0
djangorestframework==3.16.1
djangorestframework-gis==1.2.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
multidict==6.7.0
packaging==25.0
postgrest==2.22.4
propcache==0.4.1
psycopg2-binary==2.9.11
pycparser==2.23
pydantic==2.12.3
pydantic-core==2.41.4
pyjwt==2.10.1
python-dotenv==1.2.1
realtime==2.22.4
redis==7.0.1
requests==2.32.5
ruff==0.14.3
sniffio==1.3.1
sqlparse==0.5.3
storage3==2.22.4
strenum==0.4.15
supabase==2.22.4
supabase-auth==2.22.4
supabase-functions==2.22.4
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.5.0
uv==0.9.7
websockets==15.0.1
yarl==1.22.0
